refactor(profile): load member data client-side with skeleton states
Profile page now only requires a session to render; member data is
fetched async via useMember. Extracts MemberDetailsForm into its own
component and adds pulse skeletons for UPI/ID while loading.
