+++ src/app/(frontend)/profile/page.tsx
@@ -1,24 +1,10 @@
-import { headers } from "next/headers"
 import { redirect } from "next/navigation"
-import { auth } from "@/lib/auth"
+import { getSession } from "@/lib/auth-session"
 import { Routes } from "@/lib/routes"
-import { AuthService } from "@/services/auth.service"
 import { ProfilePageClient } from "./_components/ProfilePageClient"
 
-const authService = new AuthService()
-
 export default async function ProfilePage() {
-  const session = await auth.api.getSession({ headers: await headers() })
-
-  if (!session) {
-    redirect(Routes.LOGIN)
-  }
-
-  const member = await authService.getMemberFromUser(session.user)
-
-  if (!member) {
-    redirect(Routes.LOGIN)
-  }
-
-  return <ProfilePageClient member={member} />
+  const session = await getSession()
+  if (!session) redirect(Routes.LOGIN)
+  return <ProfilePageClient user={session.user} />
 }
